added authentication and authorization

+++ FastApi/TodoAPP/router/todos.py
@@ -6,12 +6,14 @@ from starlette import status
 from model.models import Todos
 from sqlalchemy.orm import Session
 from database import  db_dependency
+from .auth import get_current_user
 
 
 
 router = APIRouter()
 
 
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 
@@ -26,32 +28,49 @@ class TodoRequest(BaseModel):
   
 
 @router.get("/", status_code= status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency,
+                   db: db_dependency):
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 @router.get("/todo/{todo_id}", status_code= status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+async def read_todo(user: user_dependency, 
+                    db: db_dependency, todo_id: int = Path(gt=0)):
     
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
+    
     raise HTTPException(status_code=404, detail="Todo not found.")
 
 
 @router.post("/todo", status_code= status.HTTP_200_OK)
-async def create_todo(db: db_dependency,todo_request: TodoRequest):
+async def create_todo(user: user_dependency, 
+                      db: db_dependency,
+                      todo_request: TodoRequest):
     
-    todo_model = Todos(**todo_request.model_dump())
+    if user is None:
+        raise HTTPException(status_code=401, detail= 'Authentication Failed')
+    
+    
+    todo_model = Todos(**todo_request.model_dump(), owner_id = user.get('id'))
     
     db.add(todo_model)
     db.commit()
     
     
 @router.put("/todo/{todo_id}", status_code= status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+async def update_todo(user: user_dependency, 
+                      db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
     
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail= 'Authentication Failed')
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
         
@@ -65,12 +84,16 @@ async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int
     
     
 @router.delete("/todo/{todo_id}", status_code= status.HTTP_204_NO_CONTENT)
-async def delet_todo(db: db_dependency, todo_id:int =Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delet_todo(user: user_dependency, 
+                     db: db_dependency, todo_id:int =Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail= 'Authentication Failed')
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
         
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     
     db.commit()
